[support] Clean up mentions of things being "implementation-defined".
Remove index entries for "implementation-defined" that just point to places
where we say something is implementation-defined. This index entry is very
incomplete and we already have a separate index of such places.

Remove "implementation-defined" markers from the index of library names; they
don't make sense there.

Add an entry on the definition of NULL to the index of implementation-defined
behavior.